+++ src/components/product/ProductList/index.tsx
@@ -3,7 +3,7 @@ import * as S from './style';
 import APIErrorToast from '../../common/APIErrorToast';
 import IntersectionArea from '../../common/IntersectionArea';
 import { LoadingSpinner } from '../../common/LoadingSpinner/style';
-import { UseProductsResult } from '../../hooks/useProducts';
+import { UseProductsResult } from '../../../hooks/useProducts';
 import ProductItem from '../ProductItem';
 
 export default function ProductList({
